requirements.txt
Flask
Flask-RESTful
Flask-PyMongo
Flask-Bcrypt
Flask-JWT-Extended
marshmallow
python-dotenv